+++ lib/src/bloc_hight_light_filter.dart
@@ -1,16 +1,16 @@
 import 'package:color_observer_logger/color_observer_logger.dart';
 
-class BlocHightLightFilter {
+class BlocHighLightFilter {
   final bool colorOnly;
   final AnsiColor? color;
   final Function(String message)? onFilter;
-  BlocHightLightFilter({this.onFilter, this.colorOnly = false, this.color});
+  BlocHighLightFilter({this.onFilter, this.colorOnly = false, this.color});
   bool filter(String message) {
     return onFilter?.call(message) ?? false;
   }
 }
 
-class DefaultHighLightFilter extends BlocHightLightFilter {
+class DefaultHighLightFilter extends BlocHighLightFilter {
   DefaultHighLightFilter({
     bool colorOnly = true,
     AnsiColor? color,